\name{checknPV}
\alias{checknPV}

\title{Check the input of the experimental design function nPV.}
\description{
For internal use. Check the input of the experimental design function nPV and produce some warnings.}
\usage{
checknPV(x)
}
\arguments{
  \item{x}{an object of class \code{"nPV"}}
}

\keyword{misc}
