\name{print.BDtest}
\alias{print.BDtest}
\title{
Detailed print out for BDtest
}
\description{
Print details of the results of the function BDtest on the screen
}
\usage{
\method{print}{BDtest}(x, ...)
}

\arguments{
  \item{x}{ an object of class \code{"BDtest"}}
  \item{\dots}{ further arguments to be passed to \code{print} internally}
}

\keyword{ print }
