\name{print.nPV}
\alias{print.nPV}
\title{Detailed print out for nPV}
\description{Print details of the results of the experimental design function nPV on the screen}
\usage{
\method{print}{nPV}(x, ...)
}
\arguments{
  \item{x}{ an object of class \code{"nPV"}}
  \item{\dots}{ further arguments to be passed to \code{print}}
}
\seealso{
\code{\link{plotnPV}}, \code{\link{plotnPV2}} to plot the results of \code{nPV}
}
\keyword{ print }
